[Bug-fix] bootloader now handles exception if dfu-util is not installed on the machine

+++ owntech/scripts/pre_recovery.py
@@ -7,14 +7,6 @@ import subprocess
 Import("env")
 platformpio = env.PioPlatform()
 
-#Dummy function to print a user friendly message using env.VerboseAction() 
-#After successfully loading an example.
-def PrintSuccess(target, source, env):
-    build_dir = os.path.join(".", ".pio", "build")
-    shutil.rmtree(build_dir) #Removes build folder in order to trigger a clean compilation
-    #Necessary to make sure we go through the post action triggered by the build of firmware.bin...
-    return
-
 def installBootloader(target, source, env):
     # Make sure file is available locally
     third_party_dir = os.path.join(".", "owntech", "bootloader")
@@ -38,41 +30,56 @@ def installBootloader(target, source, env):
                         "dfu-util.exe" if platform.system() == 'Windows' else \
                         None
     
-    dfusuffix_path = [platformpio.get_package_dir("tool-dfuutil") + "/bin/" + dfu_suffix]
-    dfuutil_path = [platformpio.get_package_dir("tool-dfuutil") + "/bin/" + dfu_util]
-    
-    suffix_flags = ["-v 0x0483", \
-                    "-p 0xDF11", \
-                    "-d 0xffff", \
-                    "-a", \
-                    bootloader_path \
-                    ]
-    upload_flags = [ "-a 0", \
-                    "-s 0x08000000", \
-                    "-D", 
-                    bootloader_path \
-                    ]
-    # Sign the bootloader binary using dfu-suffix
-    suffix_command =  dfusuffix_path + suffix_flags
-    upload_command =  dfuutil_path + upload_flags
+    if platformpio.get_package_dir("tool-dfuutil") is None:
+        pkg_install = [
+                        "pio",
+                        "pkg",
+                        "install",
+                        "--global",
+                        "--tool",
+                        "platformio/tool-dfuutil"
+        ]
+        subprocess.run(pkg_install)
 
-    try:
-        # Execute the command
-        subprocess.run(suffix_command, check=True)
-    except subprocess.CalledProcessError as e:
-        print(f"Error executing mcumgr command: {e}")
+    if platformpio.get_package_dir("tool-dfuutil") is not None:
+
+        dfusuffix_path = [platformpio.get_package_dir("tool-dfuutil") + "/bin/" + dfu_suffix]
+        dfuutil_path = [platformpio.get_package_dir("tool-dfuutil") + "/bin/" + dfu_util]
+        
+        suffix_flags = ["-v 0x0483", \
+                        "-p 0xDF11", \
+                        "-d 0xffff", \
+                        "-a", \
+                        bootloader_path \
+                        ]
+        upload_flags = [ "-a 0", \
+                        "-s 0x08000000", \
+                        "-D", 
+                        bootloader_path \
+                        ]
+        # Sign the bootloader binary using dfu-suffix
+        suffix_command =  dfusuffix_path + suffix_flags
+        upload_command =  dfuutil_path + upload_flags
+
+        try:
+            # Execute the command
+            subprocess.run(suffix_command, check=True)
+        except subprocess.CalledProcessError as e:
+            print(f"Error executing mcumgr command: {e}")
+            os.remove(bootloader_path)
+            exit(-1)
+        
+        try:
+            # Execute the command
+            subprocess.run(upload_command)
+        except subprocess.CalledProcessError as e:
+            print(f"Error executing mcumgr command: {e}")
+            os.remove(bootloader_path)
+            exit(-1)
+        
         os.remove(bootloader_path)
-        exit(-1)
-    
-    try:
-        # Execute the command
-        subprocess.run(upload_command)
-    except subprocess.CalledProcessError as e:
-        print(f"Error executing mcumgr command: {e}")
-        os.remove(bootloader_path)
-        exit(-1)
-    
-    os.remove(bootloader_path)
+    else : 
+        print("Dependencies have been freshly installed, please launch action again")
 
 #Defines the targets for each example referenced in library.json and create a nice GUI.
 env.AddTarget(
